+++ src/Router.tsx
@@ -3,7 +3,7 @@ import { ROUTE } from '@/common/constants/route';
 import SelectGroup from '@/pages/selectGroup';
 import GroupNameSetup from '@/pages/groupSetup/groupNameSetup';
 import PasswordSetup from '@/pages/groupSetup/passwordSetup';
-import Create from './pages/create';
+import CreateBill from './pages/createBill';
 
 function Router() {
   return (
@@ -11,7 +11,7 @@ function Router() {
       <Route path={ROUTE.selectGroup} element={<SelectGroup />} />
       <Route path={ROUTE.groupSetupName} element={<GroupNameSetup />} />
       <Route path={ROUTE.groupSetupPassword} element={<PasswordSetup />} />
-      <Route path={ROUTE.create} element={<Create />} />
+      <Route path={ROUTE.createBill} element={<CreateBill />} />
     </Routes>
   );
 }
